Move define.TempDirForURL() to pkg/tmpdir.ForURL()
Move the generic "download a thing" function to a package that wasn't
originally intended to only include type definitions, but leave a
wrapper in the old location for compatibility's sake.
